feat: add split note chat foundation
