+++ apps/staged/src/lib/shared/viewport.svelte.ts
@@ -1,18 +1,22 @@
 const MOBILE_BREAKPOINT_PX = 768;
+const SPLIT_BREAKPOINT_PX = 1000;
 
 export const viewport = $state({
   isMobile: false,
+  canSplit: false,
   // Visual shortcut affordances only; not a hardware keyboard capability check.
   showShortcutHints: true,
 });
 
 let mediaQuery: MediaQueryList | null = null;
+let splitMediaQuery: MediaQueryList | null = null;
 let coarsePointerQuery: MediaQueryList | null = null;
 let noHoverQuery: MediaQueryList | null = null;
 let subscriberCount = 0;
 
 function syncViewport() {
   viewport.isMobile = mediaQuery?.matches ?? false;
+  viewport.canSplit = splitMediaQuery?.matches ?? false;
   viewport.showShortcutHints = !(
     (coarsePointerQuery?.matches ?? false) ||
     (noHoverQuery?.matches ?? false)
@@ -28,6 +32,10 @@ export function watchViewport(): () => void {
     mediaQuery = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT_PX}px)`);
     mediaQuery.addEventListener('change', syncViewport);
   }
+  if (!splitMediaQuery) {
+    splitMediaQuery = window.matchMedia(`(min-width: ${SPLIT_BREAKPOINT_PX}px)`);
+    splitMediaQuery.addEventListener('change', syncViewport);
+  }
   if (!coarsePointerQuery) {
     coarsePointerQuery = window.matchMedia('(pointer: coarse)');
     coarsePointerQuery.addEventListener('change', syncViewport);
@@ -42,9 +50,11 @@ export function watchViewport(): () => void {
     subscriberCount = Math.max(0, subscriberCount - 1);
     if (subscriberCount === 0) {
       mediaQuery?.removeEventListener('change', syncViewport);
+      splitMediaQuery?.removeEventListener('change', syncViewport);
       coarsePointerQuery?.removeEventListener('change', syncViewport);
       noHoverQuery?.removeEventListener('change', syncViewport);
       mediaQuery = null;
+      splitMediaQuery = null;
       coarsePointerQuery = null;
       noHoverQuery = null;
     }
